Fix GitHub bug #92. Fix fmod() ambiguity error.
(cherry picked from commit 62dcb1c94b837bf181da75b075f6e8a6a01edaaf)

+++ src/fileappender.cxx
@@ -645,7 +645,9 @@ round_time (Time const & t, time_t seconds)
 {
     return Time (
         t.getTime ()
-        - static_cast<time_t>(std::fmod (t.getTime (), seconds)));
+        - static_cast<time_t>(std::fmod (
+                static_cast<double>(t.getTime ()),
+                static_cast<double>(seconds))));
 }
